fix(cli): consolidate runCommandSilently and skip stdin pipe to avoid PowerShell deadlock
Two intertwined fixes:

1. Drop the stdin pipe in `runCommandSilently` — use `stdio: ['ignore',
   'pipe', 'pipe']` instead of `stdio: 'pipe'`. Leaving an open stdin
   pipe deadlocks any descendant `.ps1` shim whose
   `$MyInvocation.ExpectingInput` branch reads stdin to EOF before
   invoking `node` (the npm/pnpm/yarn cmd-shim pwsh template). The
   `migration-prettier-eslint-combo` snap test on Windows hung because
   `vp migrate` shells out via `runCommandSilently` to `vp dlx
   @oxlint/migrate`, and after the .cmd → PowerShell rewrite landed on
   this branch the inherited stdin pipe was never closed.

2. Consolidate the two duplicate `runCommandSilently` / `runCommand`
   pairs (`packages/cli/src/utils/command.ts` and
   `packages/cli/src/create/command.ts`) into one canonical copy in
   `utils/command.ts`. `create/command.ts` now re-exports them and
   keeps only the create-specific helpers
   (`runCommandAndDetectProjectDir`, `getPackageRunner`,
   `formatDlxCommand`, `prependToPathToEnvs`). Aligns the previously
   diverging `runCommand` return types (`number` vs `ExecutionResult`)
   on `Promise<ExecutionResult>` — utils' `runCommand` had no callers
   so the rename is safe — and prevents the stdin fix from drifting
   between two copies.
